debug: refactor debug classes
