--- src/platform/windows/gui/debug/gpu/gpu.h
@@ -1,6 +1,0 @@
-#pragma once
-#include "platform/windows/gui/gui.h"
-
-extern bool showGpuWindow;
-
-void gpuWindow(System *sys);
