+++ src/platform/windows/gui/debug/kernel.h
@@ -1,5 +1,6 @@
 #pragma once
-#include "platform/windows/gui/gui.h"
+
+struct System;
 
 extern bool showKernelWindow;
