+++ src/platform/windows/gui/debug/spu.cpp
@@ -1,11 +1,12 @@
+#include "spu.h"
 #include <fmt/core.h>
 #include <imgui.h>
 #include <vector>
 #include "device/spu/spu.h"
-#include "gui.h"
+#include "system.h"
 
+namespace gui::debug {
 using namespace spu;
-extern bool showSpuWindow;
 
 void channelsInfo(spu::SPU* spu, bool parseValues) {
     const int COL_NUM = 11;
@@ -76,7 +77,7 @@ void channelsInfo(spu::SPU* spu, bool parseValues) {
     column("Repeat addr");
     column("ADSR");
     column("Flags");
-    for (int i = 0; i < SPU::VOICE_COUNT; i++) {
+    for (int i = 0; i < spu::SPU::VOICE_COUNT; i++) {
         auto& v = spu->voices[i];
 
         if (v.state == Voice::State::Off) {
@@ -134,38 +135,40 @@ void channelsInfo(spu::SPU* spu, bool parseValues) {
 void reverbInfo(spu::SPU* spu) {
     const int width = 8;
     const int height = 4;
-    const std::array<const char*, 32> registerDescription = {{"dAPF1   - Reverb APF Offset 1",
-                                                              "dAPF2   - Reverb APF Offset 2",
-                                                              "vIIR    - Reverb Reflection Volume 1",
-                                                              "vCOMB1  - Reverb Comb Volume 1",
-                                                              "vCOMB2  - Reverb Comb Volume 2",
-                                                              "vCOMB3  - Reverb Comb Volume 3",
-                                                              "vCOMB4  - Reverb Comb Volume 4",
-                                                              "vWALL   - Reverb Reflection Volume 2",
-                                                              "vAPF1   - Reverb APF Volume 1",
-                                                              "vAPF2   - Reverb APF Volume 2",
-                                                              "mLSAME  - Reverb Same Side Reflection Address 1 Left",
-                                                              "mRSAME  - Reverb Same Side Reflection Address 1 Right",
-                                                              "mLCOMB1 - Reverb Comb Address 1 Left",
-                                                              "mRCOMB1 - Reverb Comb Address 1 Right",
-                                                              "mLCOMB2 - Reverb Comb Address 2 Left",
-                                                              "mRCOMB2 - Reverb Comb Address 2 Right",
-                                                              "dLSAME  - Reverb Same Side Reflection Address 2 Left",
-                                                              "dRSAME  - Reverb Same Side Reflection Address 2 Right",
-                                                              "mLDIFF  - Reverb Different Side Reflect Address 1 Left",
-                                                              "mRDIFF  - Reverb Different Side Reflect Address 1 Right",
-                                                              "mLCOMB3 - Reverb Comb Address 3 Left",
-                                                              "mRCOMB3 - Reverb Comb Address 3 Right",
-                                                              "mLCOMB4 - Reverb Comb Address 4 Left",
-                                                              "mRCOMB4 - Reverb Comb Address 4 Right",
-                                                              "dLDIFF  - Reverb Different Side Reflect Address 2 Left",
-                                                              "dRDIFF  - Reverb Different Side Reflect Address 2 Right",
-                                                              "mLAPF1  - Reverb APF Address 1 Left",
-                                                              "mRAPF1  - Reverb APF Address 1 Right",
-                                                              "mLAPF2  - Reverb APF Address 2 Left",
-                                                              "mRAPF2  - Reverb APF Address 2 Right",
-                                                              "vLIN    - Reverb Input Volume Left",
-                                                              "vRIN    - Reverb Input Volume Right"}};
+    const std::array<const char*, 32> registerDescription = {
+        {"dAPF1   - Reverb APF Offset 1",
+         "dAPF2   - Reverb APF Offset 2",
+         "vIIR    - Reverb Reflection Volume 1",
+         "vCOMB1  - Reverb Comb Volume 1",
+         "vCOMB2  - Reverb Comb Volume 2",
+         "vCOMB3  - Reverb Comb Volume 3",
+         "vCOMB4  - Reverb Comb Volume 4",
+         "vWALL   - Reverb Reflection Volume 2",
+         "vAPF1   - Reverb APF Volume 1",
+         "vAPF2   - Reverb APF Volume 2",
+         "mLSAME  - Reverb Same Side Reflection Address 1 Left",
+         "mRSAME  - Reverb Same Side Reflection Address 1 Right",
+         "mLCOMB1 - Reverb Comb Address 1 Left",
+         "mRCOMB1 - Reverb Comb Address 1 Right",
+         "mLCOMB2 - Reverb Comb Address 2 Left",
+         "mRCOMB2 - Reverb Comb Address 2 Right",
+         "dLSAME  - Reverb Same Side Reflection Address 2 Left",
+         "dRSAME  - Reverb Same Side Reflection Address 2 Right",
+         "mLDIFF  - Reverb Different Side Reflect Address 1 Left",
+         "mRDIFF  - Reverb Different Side Reflect Address 1 Right",
+         "mLCOMB3 - Reverb Comb Address 3 Left",
+         "mRCOMB3 - Reverb Comb Address 3 Right",
+         "mLCOMB4 - Reverb Comb Address 4 Left",
+         "mRCOMB4 - Reverb Comb Address 4 Right",
+         "dLDIFF  - Reverb Different Side Reflect Address 2 Left",
+         "dRDIFF  - Reverb Different Side Reflect Address 2 Right",
+         "mLAPF1  - Reverb APF Address 1 Left",
+         "mRAPF1  - Reverb APF Address 1 Right",
+         "mLAPF2  - Reverb APF Address 2 Left",
+         "mRAPF2  - Reverb APF Address 2 Right",
+         "vLIN    - Reverb Input Volume Left",
+         "vRIN    - Reverb Input Volume Right"},
+    };
 
     ImGui::Checkbox("Force off", &spu->forceReverbOff);
 
@@ -227,14 +230,11 @@ void debugTools(spu::SPU* spu) {
     ImGui::Checkbox("Force Pitch Modulation off", &spu->forcePitchModulationOff);
 }
 
-void spuWindow(spu::SPU* spu) {
-    if (!showSpuWindow) {
-        return;
-    }
+void SPU::spuWindow(spu::SPU* spu) {
     const auto treeFlags = ImGuiTreeNodeFlags_CollapsingHeader | ImGuiTreeNodeFlags_DefaultOpen;
     static bool parseValues = true;
 
-    ImGui::Begin("SPU", &showSpuWindow);
+    ImGui::Begin("SPU", &spuWindowOpen);
 
     if (ImGui::TreeNodeEx("Channels", treeFlags)) channelsInfo(spu, parseValues);
     if (ImGui::TreeNodeEx("Reverb", treeFlags)) reverbInfo(spu);
@@ -245,3 +245,8 @@ void spuWindow(spu::SPU* spu) {
     renderSamples(spu);
     ImGui::End();
 }
+
+void SPU::displayWindows(System* sys) {
+    if (spuWindowOpen) spuWindow(sys->spu.get());
+}
+};  // namespace gui::debug
