+++ src/platform/windows/gui/debug/spu.h
@@ -1,0 +1,17 @@
+#pragma once
+
+struct System;
+
+namespace spu {
+struct SPU;
+};
+
+namespace gui::debug {
+class SPU {
+    void spuWindow(spu::SPU* spu);
+
+   public:
+    bool spuWindowOpen = false;
+    void displayWindows(System* sys);
+};
+}  // namespace gui::debug
